refactor: remoe code duplication
